pyproject.toml
[project]
name = "visualvault"
version = "0.1.0"
description = "Smart Visual Asset Intelligence API - An ML-powered image analysis and search platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",  # For file uploads

    # Data Validation
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Database
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",  # Async PostgreSQL driver
    "psycopg2-binary>=2.9.9",  # Sync PostgreSQL driver (for Alembic)

    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",  # JWT tokens
    "passlib>=1.7.4",  # Password hashing
    "bcrypt>=4.0.0,<5.0.0",  # bcrypt backend for passlib (pinned for compatibility)
    "slowapi>=0.1.9",  # Rate limiting

    # Task Queue
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",

    # ML & Computer Vision
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.36.0",  # For CLIP
    "ultralytics>=8.0.0",  # YOLOv8 for object detection
    "pillow>=10.2.0",
    "easyocr>=1.7.0",  # OCR
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",  # For k-means clustering

    # Utilities
    "httpx>=0.26.0",  # Async HTTP client
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",  # Structured logging
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",  # For TestClient
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]
